Tweet: Mark retweeted retweets as retweeted(rly)
Let's say I'm following person A. Person A now retweets a tweet from
person B. Since we follow Person A, we now see that tweet. We retweet
it. Now, when we get the data from Twitter again, neither person A's nor
person B's tweet have the 'retweeted' property set to true, which is
weird. However, it does have the current_user_retweet is set so we can
use that to identify if we retweeted it.