TweetListEntry: Fix refcount
Using 'this' inside the callback when connecting to signals currently
increases the ref count of 'this'. This makes it impossible for
instances of TweetListEntry to freed when the window they're in is
closed. We work around this by just creating a separate method for all
our signal handlers and connecting this way.
For more information see
https://bugzilla.gnome.org/show_bug.cgi?id=624624