MainWindow: Fix scroll position 'saving'
Ok. If we set the 'active' property of the tool button of a page, the
clicked signal gets fired(which is reasonable). BUT we call switch_page
in out clicked handler. That is generally correct(for scrolling up on
second click, i.e. double_open behavior), but not if we are
already switching the page.
Introduce a window-local page_switch_lock to 'fix' this.