TweetListEntry: Remove the inline reply revealer
It was kinda weird to have 2 completely different ways to reply to a
tweet. Also, the inline reply really tried to be too much UI in too less
space -- didn't display how many characters are left, didn't preview the
inline media, etc. -- so just remove it and show the usual
ComposeTweetWindow which just directly grabs the keyboard focus so no
need to switch to the mouse.
Also makes instantiating a TweetListEntry faster.