DefaultTimeline: Fix crash
This happened whenever the user had >1 windows open, then scrolled down
in window A, scrolled up in it again. Now the tweet_remove_timeout in
DefaultTimeline was statrted. Now the user closed window A, and by the
time the tweet_remove_timeout was executed, the widgets didn't exist
anymore, causing he entire application to crash.