Add uninstall target to build system
Shamelessly taken from the CMake wiki and modified slightly to match the
style of the rest of the build files for Corebird. CMake expects
uninstallations to be handled by the package manager, but that's a pain
if you have have to install it by hand from source because your
distribution doesn't have a way of installing it with it's package
manager yet.