Don't set the title of a non-window
Activating the button of the first page in the constructor of MainWidget
was used to initialize that first page and to set the title of the
window the MainWidget is in. But at that point it didn't have a parent
yet so that caused an assertion to fail later on.