MainWindow: Reword the application handling
In certain parts of the code, we are now listening to the application's
window-added signal. In MainWindow, we used the GLib.Object () syntax to
call the base constructor with the given application (that we need to
connect to some of its signals). But this also emitted the window-added
signal which makes other parts of our code fail because at that point,
everythign in MainWindow is untouched, e.g. the account is still null.