app: Don't open account windows twice
The problematic change was introduced in 7a90917128b824f59b0d82381907.
That commit was caring about the case of outdated screen_names in the
startup-accounts array, and just opened the first configured account
when no window of any startup-accunt was opened. But that will naturally
also be the case if all configured startup-accounts have windows opened
(without any of them having an outdated screen_name).

So, just look if *any* configured account exists without an open window
and just open a window for that account and return.

Fixes #309