Tweet: Add flags for the reason of hiding it
Something we currently can't do is e.g. adding a filter that hides a
particular tweet, then unfollow the author of that tweet, then follow
them again. The previously hidden (because filtered) tweet will be shown
again, because we only have one bit of information for hiding it (i.e.
the "visible" flag in GtkWidget). Expand this to one integer and use it
as a bitfield for the various reasons there are for hiding a tweet. We
then can manipulate those flags however we want and only show the tweet
again if hidden_flags == 0.