Update TextTransform.vala
Applying text.index_of_nth_char would cause the cur_end index to 'desync' on certain japanese text with entities at the end. It woudl cause the next substring call to crash on some inputs (due to invalid indices).

I noticed this bug when trying to search for "倉庫番" in corebird. This edit stopped the crash and allows me to search that term.