MainWidget: Use the correct page_id in switch_page
if page_id is PREVIOUS or NEXT, the call to history.back/forward will
cause history.current to point to the NEW page instead of the old one.
However, we already save the current page id in current_page, so just
use that to call on_leave on the current page.