TweetInfoPage: Don't create a new Tweet if we already have one
The only data we are really updating is the via_labe, the favorite count
and the retweet count. Creating a new tweet unfortunately also downloads
the media, so we try to avoid that unless we have to (i.e. we get to
that point without a tweet instance, which is the case when the user
clicks on a tweet link.