Try to fix #457 by only removing Twitter image URLs
This is basically what Twitter does with its "cards" (Flic.kr links
remain but the card shows the image).

The only issue so far is if someone uses IFTTT or similar to auto-
post Flickr photos. IFTTT puts the Flickr URL in *and* uploads the
photo to Twitter, which results in us seeing it twice.