Restrct media button height when replying
Generally, the minimum height of a widget may not increase when the
width increases (for hfw widget). However, when inside a GtkListBoxRow,
we need exactly this behavior since the row will only ever get the
minimum height allocated. When the GtkListBox is not scrolling however,
we can't implement this without GtkWindow heavily underallocating its
contents, so use a lower minimum size and a fixed maximum height for the
media buttons instead.