Recalculate min_id on removing tweets
As per bug #685, the min_id in the model gets updated when adding tweets but never when removing. The commented out call to update_min_max_id seems to be integral to this working properly.
This needs to be done *before* the emit_items_changed in case anything that gets notified needs to know the new min_id.