text transform: Care about from/to position and display range
The old code just moving all the entities with a .from >
display_range_start to the start of the text was wrong, since that meant
that e.g. links inside the display range now potentially were before the
display_range_start, so they were simply discarded later when processing
the entities.

Instead, leave .to and .from as they are and look at the `.from -
display_range_start` and `.to - display_range_start` values instead.