CbMiniTweet: Include all mentions when no duplicate reply user was found
This happens when a tweet has a in_reply_to user attached, but the user
mentions don't include that user, at leat not before the display_range
starts. In that case, we still need to count the all the user mentions,
so we can't start parsing them at n_reply_users since that will be 1.