MainWindow: Show the menu button after creating it
widgets are still invisible by default in gtk3 and in the case of a
MainWindow instance created without a valid Account instance, the
following show_all call does not affect the app menu button since it
will only be crated when the account is changed to a valid one. So, make
sure we show the button we just created.

Fixes #738